backend/mutual-funds-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance==0.2.36
pandas==2.2.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
